% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat2}
\alias{dat2}
\title{dat2}
\format{An object of class \code{rseHandle} of length 1.}
\usage{
dat2
}
\description{
A small experiment containing 100 peaks x 50 samples
colData have some random attributes and the peaks
are fixed from chr12.
}
\details{
Includes a .bed file of row data, a .txt of col
data, a .csv of the counts, a .sqlite dataframe
of the individual data that have been processed,
and a .rds of the object read into an S4 object.
}
\keyword{datasets}

