% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvalues.R
\docType{methods}
\name{getvalues}
\alias{getvalues}
\alias{getvalues,rseHandle-method}
\title{Evaluate an \code{rseHandle} object and return a
\code{RangedSummarizedExperiment}}
\usage{
getvalues(handle, format = "sparse")

\S4method{getvalues}{rseHandle}(handle, format = "sparse")
}
\arguments{
\item{handle}{An \code{rseHandle} object.}

\item{format}{= "sparse" or "normal". Representation of data
when evaluted.}
}
\value{
Returns a \code{RangedSummarizedExperiment} object.
}
\description{
\code{getvalues} evaluates any valid  \code{rseHandle} object
and loads data into memory that is ordinarily located on disk.
}

