% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefinitions.R
\docType{class}
\name{rseHandle-class}
\alias{rseHandle}
\alias{rseHandle-class}
\title{rseHandle}
\description{
A class to represent  matrices stored in HDF5
and sqlite files where the column and row annotations
are wrapped in a \code{RangedSummarizedExperiment}
class.
}
\section{Slots}{

\describe{
\item{\code{rowMap}}{Named vector mapping indicies of
the current rows to indicies in the original file.}

\item{\code{colMap}}{Named vector mapping indicies of
the current columns to the indicies in the original file.}
}}

