% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\docType{methods}
\name{rseHandleSubset}
\alias{[,rseHandle,logical,logical,missing-method}
\alias{[,rseHandle,logical,missing,missing-method}
\alias{[,rseHandle,missing,logical,missing-method}
\alias{[,rseHandle,missing,numeric,missing-method}
\alias{[,rseHandle,numeric,missing,missing-method}
\alias{[,rseHandle,numeric,numeric,missing-method}
\alias{rseHandleSubset}
\alias{rseHandleSubset,rseHandle,logical,logical-method}
\alias{rseHandleSubset,rseHandle,logical,missing-method}
\alias{rseHandleSubset,rseHandle,missing,logical-method}
\alias{rseHandleSubset,rseHandle,missing,numeric-method}
\alias{rseHandleSubset,rseHandle,numeric,missing-method}
\alias{rseHandleSubset,rseHandle,numeric,numeric-method}
\title{Subset an \code{rseHandle} object}
\usage{
rseHandleSubset(x, i, j)

\S4method{rseHandleSubset}{rseHandle,numeric,numeric}(x, i, j)

\S4method{rseHandleSubset}{rseHandle,missing,numeric}(x, i, j)

\S4method{rseHandleSubset}{rseHandle,numeric,missing}(x, i, j)

\S4method{rseHandleSubset}{rseHandle,missing,logical}(x, i, j)

\S4method{rseHandleSubset}{rseHandle,logical,missing}(x, i, j)

\S4method{rseHandleSubset}{rseHandle,logical,logical}(x, i, j)

\S4method{[}{rseHandle,numeric,numeric,missing}(x, i, j, drop)

\S4method{[}{rseHandle,missing,numeric,missing}(x, i, j, drop)

\S4method{[}{rseHandle,numeric,missing,missing}(x, i, j, drop)

\S4method{[}{rseHandle,logical,logical,missing}(x, i, j, drop)

\S4method{[}{rseHandle,missing,logical,missing}(x, i, j, drop)

\S4method{[}{rseHandle,logical,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{An \code{rseHandle} object to be subset.}

\item{i}{Numeric indices to be subset from rows.}

\item{j}{Numeric indices to be subset from columns}
}
\value{
Returns a subsetted \code{sqlSparseHandle} object.
}
\description{
\code{rseHandleSubset} subsets the \code{rowData} or \code{colData}
based on the indicies of i (rowData) and j (colData). Can
also use the `[` operator as is common in R object subsetting.
}
\examples{

}

