% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefinitions.R
\docType{class}
\name{sqlSparseHandle-class}
\alias{sqlSparseHandle}
\alias{sqlSparseHandle-class}
\title{sqlSparseHandle}
\description{
A class to represent sparse matrices stored in
and sql file where the column and row annotations
are wrapped in a \code{RangedSummarizedExperiment}
class.
}
\section{Slots}{

\describe{
\item{\code{rowMap}}{Named vector mapping indicies of
the current rows to indicies in the original file.}

\item{\code{colMap}}{Named vector mapping indicies of
the current columns to the indicies in the original file.}
}}

