% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\docType{methods}
\name{subsetByOverlaps,rseHandle,GRanges-method}
\alias{subsetByOverlaps,rseHandle,GRanges-method}
\title{Subset an \code{rseHandle} object by region defined in a 
\code{GRanges} object}
\usage{
\S4method{subsetByOverlaps}{rseHandle,GRanges}(query, subject, maxgap = 0L,
  minoverlap = 1L, type = c("any", "start", "end", "within", "equal"),
  invert = FALSE, ...)
}
\arguments{
\item{query}{An \code{rseHandle} object}

\item{subject}{A \code{GRanges} object of the regions of intrest}

\item{maxgap}{Refer to \code{GRanges} documentation of this function.}

\item{minoverlap}{Refer to \code{GRanges} documentation of this function.}

\item{type}{Refer to \code{GRanges} documentation of this function.}

\item{invert}{Refer to \code{GRanges} documentation of this function.}

\item{...}{Refer to \code{GRanges} documentation of this function.}
}
\value{
A \code{rseHandle} object subsetted
after the subjet was filtered
}
\description{
Subset an \code{rseHandle} object by region defined in a 
\code{GRanges} object
}
\examples{
library(GenomicRanges)
chr12reg <- GRanges(seqnames=c("chr12"),
     ranges = IRanges(start=c(56150000),end=c(56500000)))
dat2 <- readRDS(system.file("rds", "dat2_ssh.rds", package = "yolo"))
ss <- subsetByOverlaps(dat2, chr12reg)

}

