% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo.R
\docType{package}
\name{yolo}
\alias{yolo}
\alias{yolo-package}
\title{yolo: You Only Lookup One (small part of an otherwise large matrix)}
\description{
The yolo package contains a suite of methods and S4 data objects to
efficiently manipulate large genomic experiments in the 
\code{(Ranged)SummarizedExperiment} inheritence scheme for 
manipulating data.
}
\section{yolo classes}{
 Listed in classDefinitions.R. These are
primarily implementations to keep indicies ordered in sparse
and regular matrices using sqlite and HDF5.
}

