% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\docType{methods}
\name{yoloHandleSubset}
\alias{[,yoloHandle,logical,logical,missing-method}
\alias{[,yoloHandle,logical,missing,missing-method}
\alias{[,yoloHandle,missing,logical,missing-method}
\alias{[,yoloHandle,missing,numeric,missing-method}
\alias{[,yoloHandle,numeric,missing,missing-method}
\alias{[,yoloHandle,numeric,numeric,missing-method}
\alias{yoloHandleSubset}
\alias{yoloHandleSubset,yoloHandle,logical,logical-method}
\alias{yoloHandleSubset,yoloHandle,logical,missing-method}
\alias{yoloHandleSubset,yoloHandle,missing,logical-method}
\alias{yoloHandleSubset,yoloHandle,missing,numeric-method}
\alias{yoloHandleSubset,yoloHandle,numeric,missing-method}
\alias{yoloHandleSubset,yoloHandle,numeric,numeric-method}
\title{Subset a \code{yoloHandle} object}
\usage{
yoloHandleSubset(x, i, j)

\S4method{yoloHandleSubset}{yoloHandle,numeric,numeric}(x, i, j)

\S4method{yoloHandleSubset}{yoloHandle,missing,numeric}(x, i, j)

\S4method{yoloHandleSubset}{yoloHandle,numeric,missing}(x, i, j)

\S4method{yoloHandleSubset}{yoloHandle,missing,logical}(x, i, j)

\S4method{yoloHandleSubset}{yoloHandle,logical,missing}(x, i, j)

\S4method{yoloHandleSubset}{yoloHandle,logical,logical}(x, i, j)

\S4method{[}{yoloHandle,numeric,numeric,missing}(x, i, j, drop)

\S4method{[}{yoloHandle,missing,numeric,missing}(x, i, j, drop)

\S4method{[}{yoloHandle,numeric,missing,missing}(x, i, j, drop)

\S4method{[}{yoloHandle,logical,logical,missing}(x, i, j, drop)

\S4method{[}{yoloHandle,missing,logical,missing}(x, i, j, drop)

\S4method{[}{yoloHandle,logical,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A \code{yoloHandle} object to be subset.}

\item{i}{Numeric indices to be subset from rows.}

\item{j}{Numeric indices to be subset from columns}

\item{drop}{Other argument to pass to [}
}
\value{
Returns a subsetted \code{yoloHandle} object.
}
\description{
\code{yoloHandleSubset} subsets the \code{rowData} or \code{colData}
based on the indicies of i (rowData) and j (colData). Can
also use the `[` operator as is common in R object subsetting.
}
\examples{
dat2 <- readRDS(system.file("rds", "dat2_ssh.rds", package = "yolo"))
d <- dat2[1:5, 3:4]
d <- yoloHandleSubset(dat2, 1:5, 3:4)

}

