% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvalues.R
\docType{methods}
\name{getvalues}
\alias{getvalues}
\alias{getvalues,yoloHandle-method}
\title{Evaluate an \code{rseHandle} object and return a
\code{RangedSummarizedExperiment} or a \code{seHandle}
object and rturn a \code{SummarizedExperiment}}
\usage{
getvalues(handle, format = "sparse")

\S4method{getvalues}{yoloHandle}(handle, format = "sparse")
}
\arguments{
\item{handle}{An  \code{seHandle} or \code{rseHandle} object.}

\item{format}{= "sparse" or "normal". Representation of data
when evaluted.}
}
\value{
Returns a \code{(Ranged)SummarizedExperiment} object.
}
\description{
\code{getvalues} evaluates any valid  \code{yoloHandle} object
and loads data into memory that is ordinarily located on disk.
}

