% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\docType{methods}
\name{subsetByOverlaps,rseHandle,GRanges-method}
\alias{subsetByOverlaps,rseHandle,GRanges-method}
\title{Subset an \code{rseHandle} object by region defined in a 
\code{GRanges} object}
\usage{
\S4method{subsetByOverlaps}{rseHandle,GRanges}(query, subject, maxgap = 0L,
  minoverlap = 1L, type = c("any", "start", "end", "within", "equal"), ...)
}
\arguments{
\item{query}{An \code{rseHandle} object}

\item{subject}{A \code{GRanges} object of the regions of intrest}
}
\value{
A \code{rseHandle} object subsetted
after the subjet was filtered
}
\description{
Subset an \code{rseHandle} object by region defined in a 
\code{GRanges} object
}

